Added authentication checks in every controller method

+++ src/main/java/com/paytmbank/middleware/emgs/controller/EmployeeController.java
@@ -2,10 +2,12 @@ package com.paytmbank.middleware.emgs.controller;
 
 import com.paytmbank.middleware.emgs.details.EmployeeDetailsBasic;
 import com.paytmbank.middleware.emgs.entity.Employee;
+import com.paytmbank.middleware.emgs.exception.EmployeeNotFound;
 import com.paytmbank.middleware.emgs.exception.UnauthorizedUser;
 import com.paytmbank.middleware.emgs.security.UserSecurity;
 import com.paytmbank.middleware.emgs.service.EmployeeService;
 
+import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
@@ -18,6 +20,12 @@ import org.springframework.web.bind.annotation.*;
 public class EmployeeController {
     @Autowired
     EmployeeService employeeService;
+
+    Employee currentUser() throws EmployeeNotFound {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        UserSecurity user = (UserSecurity) auth.getPrincipal();
+        return employeeService.find(user.getUsername());
+    }
 
     /* Get the list of all users. This controller method is strictly for testing purposes. */
     /* Although this will be available to the super_admin during production (if any such thing happens). */
@@ -33,10 +41,13 @@ public class EmployeeController {
         /* Basic returnable object, detailing the error and status. */
         EmployeeDetailsBasic obj = new EmployeeDetailsBasic();
         try {
-            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-            String role = auth.getAuthorities().toArray()[0].toString();
-            if (role != "ROLE_ADMIN") throw new UnauthorizedUser("You must be admin to perform this operation.");
+            // Check if the employee who is creating is an admin.
+            Employee currentUser = currentUser();
 
+            if (currentUser.getRole() != "ROLE_ADMIN")
+                throw new UnauthorizedUser("You must be admin to perform this operation.");
+
+            // Create method uses emp, do some checks and then save it to the repo.
             employeeService.create(emp);
         } catch (UnauthorizedUser e) {
             obj.setStatus("Failure!");
@@ -60,8 +71,18 @@ public class EmployeeController {
     public ResponseEntity<?> getEmployee(@PathVariable(name = "id", required = true) String id) {
         EmployeeDetailsBasic obj = new EmployeeDetailsBasic();
         try {
+            /* Check if the user who is making the request is having the same eid, or is admin or not. */
+            Employee currentUser = currentUser();
+            if (currentUser.getEid() != id && currentUser.getRole() != "ROLE_ADMIN")
+                throw new UnauthorizedUser("You are not authorized for this request.");
+
             Employee emp = employeeService.find(id);
             return ResponseEntity.ok().body(emp);
+        } catch(UnauthorizedUser e) {
+            obj.setStatus("Failure!");
+            obj.setErrorDesc(e.getLocalizedMessage());
+
+            return ResponseEntity.status(401).body(obj);
         } catch(Exception e) {
             obj.setStatus("Failure!");
             obj.setErrorDesc(e.getLocalizedMessage());
@@ -72,11 +93,21 @@ public class EmployeeController {
 
     /* Get Employee by its phone number. The phone parameter maps to Employee.phone. */
     @GetMapping("/getEmployeeByPhone/{phone}")
-    public ResponseEntity<?> getEmployeeByPhone(@PathVariable(name = "phone", required = true) String id) {
+    public ResponseEntity<?> getEmployeeByPhone(@PathVariable(name = "phone", required = true) String phone) {
         EmployeeDetailsBasic obj = new EmployeeDetailsBasic();
         try {
-            Employee emp = employeeService.findByPhone(id);
+            /* Check if the user who is making the request is having the same phone, or is admin or not. */
+            Employee currentUser = currentUser();
+            if (currentUser.getPhone() != phone && currentUser.getRole() != "ROLE_ADMIN")
+                throw new UnauthorizedUser("You are not authorized for this request.");
+
+            Employee emp = employeeService.findByPhone(phone);
             return ResponseEntity.ok().body(emp);
+        } catch(UnauthorizedUser e) {
+            obj.setStatus("Failure!");
+            obj.setErrorDesc(e.getLocalizedMessage());
+
+            return ResponseEntity.status(401).body(obj);
         } catch(Exception e) {
             obj.setStatus("Failure!");
             obj.setErrorDesc(e.getLocalizedMessage());
@@ -87,11 +118,21 @@ public class EmployeeController {
 
     /* Get Employee by its email. The email parameter maps to Employee.email. */
     @GetMapping("/getEmployeeByEmail/{email}")
-    public ResponseEntity<?> getEmployeeByEmail(@PathVariable(name = "email", required = true) String id) {
+    public ResponseEntity<?> getEmployeeByEmail(@PathVariable(name = "email", required = true) String email) {
         EmployeeDetailsBasic obj = new EmployeeDetailsBasic();
         try {
-            Employee emp = employeeService.findByEmail(id);
+            /* Check if the user who is making the request is having the same email, or is admin or not. */
+            Employee currentUser = currentUser();
+            if (currentUser.getEmail() != email && currentUser.getRole() != "ROLE_ADMIN")
+                throw new UnauthorizedUser("You are not authorized for this request.");
+
+            Employee emp = employeeService.findByEmail(email);
             return ResponseEntity.ok().body(emp);
+        } catch(UnauthorizedUser e) {
+            obj.setStatus("Failure!");
+            obj.setErrorDesc(e.getLocalizedMessage());
+
+            return ResponseEntity.status(401).body(obj);
         } catch(Exception e) {
             obj.setStatus("Failure!");
             obj.setErrorDesc(e.getLocalizedMessage());
@@ -105,8 +146,18 @@ public class EmployeeController {
     public ResponseEntity<?> delete(@PathVariable(name = "eid", required = true) String eid) {
         EmployeeDetailsBasic obj = new EmployeeDetailsBasic();
         try {
+            /* Check if the user who is making the request is having the same email, or is admin or not. */
+            Employee currentUser = currentUser();
+            if (currentUser.getEid() != eid && currentUser.getRole() != "ROLE_ADMIN")
+                throw new UnauthorizedUser("You are not authorized for this request.");
+
             employeeService.delete(eid);
             return ResponseEntity.ok().body("Delete successful.");
+        } catch(UnauthorizedUser e) {
+            obj.setStatus("Failure!");
+            obj.setErrorDesc(e.getLocalizedMessage());
+
+            return ResponseEntity.status(401).body(obj);
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(e.getLocalizedMessage());
         }
@@ -117,9 +168,19 @@ public class EmployeeController {
     public ResponseEntity<?> update(@RequestBody Employee emp) {
         EmployeeDetailsBasic obj = new EmployeeDetailsBasic();
         try {
+            /* Check if the user who is making the request is having the same eid, or is admin or not. */
+            Employee currentUser = currentUser();
+            if (currentUser.getEid() != emp.getEid() && currentUser.getRole() != "ROLE_ADMIN")
+                throw new UnauthorizedUser("You are not authorized for this request.");
+
             employeeService.update(emp);
             obj = new EmployeeDetailsBasic(emp.getEid(), emp.getFname(), emp.getSname());
             return ResponseEntity.ok().body(obj);
+        } catch(UnauthorizedUser e) {
+            obj.setStatus("Failure!");
+            obj.setErrorDesc(e.getLocalizedMessage());
+
+            return ResponseEntity.status(401).body(obj);
         } catch (Exception e) {
             obj.setStatus("Failure!");
             obj.setErrorDesc(e.getLocalizedMessage());
@@ -132,11 +193,21 @@ public class EmployeeController {
     public ResponseEntity<?> dropProject(@PathVariable(name = "eid", required = true) String eid) {
         EmployeeDetailsBasic obj = new EmployeeDetailsBasic();
         try {
+            /* Check if the user who is making the request is having the same eid, or is admin or not. */
+            Employee currentUser = currentUser();
+            if (currentUser.getEid() != eid && currentUser.getRole() != "ROLE_ADMIN")
+                throw new UnauthorizedUser("You are not authorized for this request.");
+
             employeeService.dropProject(eid);
             Employee emp = employeeService.find(eid);
             obj = new EmployeeDetailsBasic(emp.getEid(), emp.getFname(), emp.getSname());
             return ResponseEntity.ok().body(obj);
-        } catch (Exception e) {
+        } catch(UnauthorizedUser e) {
+            obj.setStatus("Failure!");
+            obj.setErrorDesc(e.getLocalizedMessage());
+
+            return ResponseEntity.status(401).body(obj);
+        }  catch (Exception e) {
             obj.setStatus("Failure!");
             obj.setErrorDesc(e.getLocalizedMessage());
             return ResponseEntity.badRequest().body(obj);
@@ -144,18 +215,30 @@ public class EmployeeController {
     }
 
     /* Grant leave to the employee
-    * Pass JSON data to this post method.
-    * The JSON data must contain these fields :
-    * "eid": {The employee Id}
-    * "lid": {The Leave Id (long type) } */
+     * Pass JSON data to this post method.
+     * The JSON data must contain these fields :
+     * "eid": {The employee Id}
+     * "lid": {The Leave Id (long type) } */
     @PostMapping("/grantLeave")
     public ResponseEntity<?> grantLeave(@RequestBody String data) {
         EmployeeDetailsBasic obj = new EmployeeDetailsBasic();
         try {
+            /* Check if the user who is making the request is having the same eid, or is admin or not. */
+            JSONObject jsonObject = new JSONObject(data);
+            String eid = jsonObject.getString("eid");
+            Employee currentUser = currentUser();
+            if (currentUser.getEid() != eid && currentUser.getRole() != "ROLE_ADMIN")
+                throw new UnauthorizedUser("You are not authorized for this request.");
+
             Employee emp = employeeService.grantLeave(data);
             obj = new EmployeeDetailsBasic(emp.getEid(), emp.getFname(), emp.getSname());
             return ResponseEntity.ok().body(obj);
-        } catch (Exception e) {
+        } catch(UnauthorizedUser e) {
+            obj.setStatus("Failure!");
+            obj.setErrorDesc(e.getLocalizedMessage());
+
+            return ResponseEntity.status(401).body(obj);
+        }  catch (Exception e) {
             obj.setStatus("Failure");
             obj.setErrorDesc(e.getLocalizedMessage());
             return ResponseEntity.badRequest().body(obj);
@@ -163,17 +246,29 @@ public class EmployeeController {
     }
 
     /* Add project to Employee
-    * Pass JSON data to this post method.
-    * Must contain two fields :
-    * "eid": {The Employee Id}
-    * "pid": {The Project Id} */
+     * Pass JSON data to this post method.
+     * Must contain two fields :
+     * "eid": {The Employee Id}
+     * "pid": {The Project Id} */
     @PostMapping("/addProject/")
     public ResponseEntity<?> addProject(@RequestBody String data) {
         EmployeeDetailsBasic obj = new EmployeeDetailsBasic();
         try {
+            /* Check if the user who is making the request is having the same eid, or is admin or not. */
+            JSONObject jsonObject = new JSONObject(data);
+            String eid = jsonObject.getString("eid");
+            Employee currentUser = currentUser();
+            if (currentUser.getEid() != eid && currentUser.getRole() != "ROLE_ADMIN")
+                throw new UnauthorizedUser("You are not authorized for this request.");
+
             Employee emp = employeeService.addProject(data);
             obj = new EmployeeDetailsBasic(emp.getEid(), emp.getFname(), emp.getSname());
             return ResponseEntity.ok().body(obj);
+        } catch(UnauthorizedUser e) {
+            obj.setStatus("Failure!");
+            obj.setErrorDesc(e.getLocalizedMessage());
+
+            return ResponseEntity.status(401).body(obj);
         } catch (Exception e) {
             obj.setStatus("Failure!");
             obj.setErrorDesc(e.getLocalizedMessage());
